New version

+++ src/main/java/org/cryptimeleon/math/serialization/BigIntegerRepresentation.java
@@ -21,7 +21,7 @@ public class BigIntegerRepresentation extends Representation {
         this.n = n;
     }
 
-    public BigIntegerRepresentation(int n) {
+    public BigIntegerRepresentation(long n) {
         this.n = BigInteger.valueOf(n);
     }
